Adding Recorded Future feed
https://go.recordedfuture.com/ukraine-ioc-list

"To help identify and analyze the relevant cyber threats, Insikt Group analysts have made public a list of Indicators of Compromise (IOCs) associated with threat actors and malware related to the Russian cyber actions against Ukraine. This list contains indicators determined by Recorded Future Insikt Group's research and will be updated regularly as new indicators are identified. This list is not intended for automated blocking but to inform detection and response efforts."